+++ vite.config.js
@@ -13,11 +13,11 @@ export default defineConfig({
                 // adminlte 4.1.0 necessary to copy and modify this file
                 // copy resources/adminlte.esm.js to bt-adminlte.esm.js and edit to comment out
                 // line 901 this._applyTheme(theme); to disable ColorMode
-                {
-                    src: 'node_modules/admin-lte/dist/js/adminlte.esm.js',
-                    dest: '../../resources/js',
-                    rename: { stripBase: true },
-                },
+                // {
+                //     src: 'node_modules/admin-lte/dist/js/adminlte.esm.js',
+                //     dest: '../../resources/js',
+                //     rename: { stripBase: true },
+                // },
                 {
                     src: 'node_modules/tom-select/dist/css/*',
                     dest: '../plugins/tom-select/css',