+++ app/Modules/Documentation/Views/documentation/Release-Notes.md
@@ -5,13 +5,16 @@ Release Notes
 ### BillingTrack 8.0.0
 - update to Laravel 13, requires PHP >= 8.3
 - update to Livewire V4
-- Adminlte Release v4
+- Adminlte Release v4.1.0
+- For Adminlte 4.1.0 to disable auto theming over BT theming, had to add to bootstrap.js <<import { ColorMode } from 'admin-lte'>><<ColorMode.prototype._applyTheme = () => {}>> , _look to remove this and replace with html data-lte-color-mode="off" in v4.1.1 in master.blade_
 - update to allow MySql DB_STRICT=true (now required for BillingTrack 8, change in .env)
 - srmklive/paypal changed maintainer to blendbyte/laravel-paypal, upgrade to 3.1 - https://github.com/blendbyte/laravel-paypal
 - update square api to ^45 (using legacy client)
 - remove ancient wkhtmltopdf driver - only dompdf now
 - cherry picked items from PR#23 (arabic language, dark mode, some core fixes)
-- fix missing url_key on VERY old clients, documents and attachments
+- NO ACTIVITY on rappasoft/laravel-livewire-tables - forked and applied PR for Laravel13/livewireV4 support and added to composer.  
+  https://github.com/cytech/laravel-livewire-tables/tree/feature/l-13-lw-4-llt380
+- Fix missing url_key on VERY old clients, documents and attachments
 
 ### BillingTrack 7.0.6
 - fix job end_time validation in workorders